New Feature 16/04

+++ src/main/java/appetito/apicardapio/controller/UsuarioDashboardController.java
@@ -3,6 +3,7 @@ package appetito.apicardapio.controller;
 import appetito.apicardapio.dto.cadastro.UsuarioDashboardCadastro;
 import appetito.apicardapio.dto.detalhamento.UsuarioDashboardDetalhamento;
 import appetito.apicardapio.dto.GetAll.UsuarioDados;
+import appetito.apicardapio.entity.Cliente;
 import appetito.apicardapio.entity.UsuarioDashboard;
 import appetito.apicardapio.repository.EstabelecimentoRepository;
 import appetito.apicardapio.repository.UsuarioDashboardRepository;
@@ -10,6 +11,7 @@ import appetito.apicardapio.security.DiscordAlert;
 import appetito.apicardapio.service.UsuarioDashboardService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpHeaders;
@@ -39,13 +41,6 @@ public class UsuarioDashboardController {
         this.usuarioService = usuarioService;
     }
 
-    @GetMapping
-    @Transactional
-    public ResponseEntity<List<UsuarioDados>> listarUsuarios() {
-        var lista = usuarioRepository.findAll().stream().map(UsuarioDados::new).toList();
-        return ResponseEntity.ok(lista);
-    }
-
     @PostMapping("/cadastrar")
     public ResponseEntity<?> cadastrarUsuarioDashboard(
             @RequestBody @Valid UsuarioDashboardCadastro dadosUsuario,
@@ -68,33 +63,50 @@ public class UsuarioDashboardController {
         return ResponseEntity.created(uri).body(new UsuarioDashboardDetalhamento(usuario));
     }
 
-  //  @Operation(summary = "Upload da imagem de perfil do usuário")
-    //@PostMapping(value = "/{id}/upload-imagem", consumes = "multipart/form-data")
-  //  public ResponseEntity<String> uploadImagemPerfil(@PathVariable Long id, @RequestPart("file") MultipartFile file) {
-  //      if (file.isEmpty()) {
-   //         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Arquivo de imagem não pode estar vazio!");
-   //     }
-//
-     //   try {
-  //          UsuarioDashboard usuario = usuarioService.salvarImagemPerfil(id, file);
-    //        return usuario != null
-    //                ? ResponseEntity.ok("Imagem de perfil salva com sucesso!")
-      //              : ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário não encontrado.");
-      //  } catch (IOException e) {
-     //       return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Erro ao salvar imagem: " + e.getMessage());
-    //   }
-  //  }
+    @PostMapping(value = "/{id}/upload-imagem", consumes = "multipart/form-data")
+    public ResponseEntity<String> uploadImagemPerfil(@PathVariable Long id, @RequestPart("file") MultipartFile file, HttpServletRequest request) {
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication.getPrincipal() instanceof UsuarioDashboard usuario)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Usuário não autenticado.");
+        }
 
-   // @GetMapping("/{id}/imagem-perfil")
-   // public ResponseEntity<byte[]> buscarImagemPerfil(@PathVariable Long id) {
-   //     byte[] imagem = usuarioService.obterImagemPerfil(id);
-    //    if (imagem == null) {
-    //        return ResponseEntity.notFound().build();
-   //     }
-    //    HttpHeaders headers = new HttpHeaders();
-    //    headers.add("Content-Type", "image/png");
-     //   return new ResponseEntity<>(imagem, headers, HttpStatus.OK);
-  //  }
+        if (!usuario.getUsuario_dashboard_id().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Você não tem permissão para atualizar a imagem de outro usuário.");
+        }
+
+        String filename = file.getOriginalFilename();
+        if (filename == null || !(filename.endsWith(".jpg") || filename.endsWith(".jpeg") || filename.endsWith(".png"))) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Arquivo de imagem inválido. Apenas .jpg, .jpeg e .png são permitidos.");
+        }
+
+        if (file.getSize() > 2 * 1024 * 1024) { // 2MB = 2 * 1024 * 1024 bytes
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("O arquivo é muito grande. O tamanho máximo permitido é 2MB.");
+        }
+
+        if (file.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Arquivo de imagem não pode estar vazio!");
+        }
+
+        try {
+            UsuarioDashboard usuarioAtualizado = usuarioService.salvarImagemPerfil(id, file, request);
+            return usuarioAtualizado != null
+                    ? ResponseEntity.ok("Imagem de perfil salva com sucesso!")
+                    : ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário não encontrado.");
+        } catch (IOException e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Erro ao salvar imagem: " + e.getMessage());
+        }
+    }
+
+    @GetMapping("/{id}/imagem-perfil")
+    public ResponseEntity<byte[]> buscarImagemPerfil(@PathVariable Long id, HttpServletRequest request) {
+        byte[] imagem = usuarioService.obterImagemPerfil(id, request);
+        if (imagem == null) {
+               return ResponseEntity.notFound().build();
+        }
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Content-Type", "image/png");
+        return new ResponseEntity<>(imagem, headers, HttpStatus.OK);
+    }
 
     @GetMapping("/me")
     public ResponseEntity<?> meuPerfil() {
@@ -106,28 +118,34 @@ public class UsuarioDashboardController {
         return ResponseEntity.ok(new UsuarioDashboardDetalhamento(usuario));
     }
 
-   // @GetMapping("/me/imagem")
-   // public ResponseEntity<byte[]> minhaImagem() {
-    //    var authentication = SecurityContextHolder.getContext().getAuthentication();
-    //    if (!(authentication.getPrincipal() instanceof UsuarioDashboard usuario)) {
-   //         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
-    //    }
-//
-   //     byte[] imagem = usuarioService.obterImagemPerfil(usuario.getUsuario_dashboard_id());
-//
-   //     return ResponseEntity.ok()
-    //            .contentType(MediaType.IMAGE_PNG)
-  ///              .body(imagem);
-   // }
+    @GetMapping("/me/imagem")
+    public ResponseEntity<byte[]> minhaImagem(HttpServletRequest request) {
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication.getPrincipal() instanceof UsuarioDashboard usuario)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<UsuarioDashboardDetalhamento> deletarUsuario(@PathVariable Long id) {
-        var usuario = usuarioRepository.findById(id).orElse(null);
+        byte[] imagem = usuarioService.obterImagemPerfil(usuario.getUsuario_dashboard_id(), request);
+
+        return ResponseEntity.ok()
+                .contentType(MediaType.IMAGE_PNG)
+                .body(imagem);
+    }
+
+    @DeleteMapping("/me")
+    public ResponseEntity<UsuarioDashboardDetalhamento> deletarUsuario() {
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if (!(authentication.getPrincipal() instanceof Cliente cliente)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
+        var usuario = usuarioRepository.findById(cliente.getId()).orElse(null);
         if (usuario == null) {
             return ResponseEntity.notFound().build();
         }
         usuarioRepository.delete(usuario);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.ok(new UsuarioDashboardDetalhamento(usuario));
     }
 
     @PutMapping("/me")
@@ -138,16 +156,19 @@ public class UsuarioDashboardController {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Usuário não autenticado.");
         }
 
+        if (!usuario.getUsuario_dashboard_id().equals(authentication.getName())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Você não tem permissão para atualizar o perfil de outro usuário.");
+        }
+
         usuario.setNome_completo(dadosAtualizados.nome_completo());
         usuario.setEmail(dadosAtualizados.email());
-      //  usuario.setIdioma_padrao(dadosAtualizados.idioma_padrao());
+        // usuario.setIdioma_padrao(dadosAtualizados.idioma_padrao()); // vou arrumar isso aqui depois
 
         if (dadosAtualizados.senha() != null && !dadosAtualizados.senha().isEmpty()) {
             usuario.setSenha(passwordEncoder.encode(dadosAtualizados.senha()));
         }
-
         usuarioRepository.save(usuario);
-
         return ResponseEntity.ok(new UsuarioDashboardDetalhamento(usuario));
     }
+
 }
